+++ packages/phrases/src/locales/it/translation/admin-console/application-details.ts
@@ -41,6 +41,9 @@ const application_details = {
   enable_admin_access: "Abilita l'accesso amministratore",
   enable_admin_access_label:
     "Abilita o disabilita l'accesso all'API di gestione. Una volta abilitato, puoi utilizzare i token di accesso per chiamare l'API di gestione a nome di questa applicazione.",
+  always_issue_refresh_token: 'Rilascia sempre il token di aggiornamento',
+  always_issue_refresh_token_label:
+    'Abilitando questa configurazione, Logto potrà rilasciare sempre token di aggiornamento, anche se `prompt=consent` non viene presentata nella richiesta di autenticazione. Tuttavia, questa pratica è scoraggiata a meno che non sia necessaria, in quanto non è compatibile con OpenID Connect e potrebbe potenzialmente causare problemi.',
   delete_description:
     "Questa azione non può essere annullata. Eliminerà definitivamente l'applicazione. Inserisci il nome dell'applicazione <span>{{name}}</span> per confermare.",
   enter_your_application_name: 'Inserisci il nome della tua applicazione',