+++ packages/phrases/src/locales/zh-tw/translation/admin-console/application-details.ts
@@ -39,6 +39,9 @@ const application_details = {
   enable_admin_access: '啟用管理訪問',
   enable_admin_access_label:
     '啟用或禁用對管理 API 的訪問。啟用後，你可以使用訪問令牌代表該應用程式調用管理 API。',
+  always_issue_refresh_token: '始終發放 Refresh Token',
+  always_issue_refresh_token_label:
+    '啟用此配置將使 Logto 無論在驗證請求中是否提供 prompt=consent，都能始終發放 Refresh Token。然而，除非必要，否則不鼓勵這種做法，因為它與 OpenID Connect 不相容並可能引起問題。',
   delete_description:
     '本操作會永久性地刪除該應用程式，且不可撤銷。輸入 <span>{{name}}</span> 確認。',
   enter_your_application_name: '輸入你的應用程式姓名',